Remove duplicate code. Split in pegasus by data type
