--- bnm/recon/daxgen/__main__.py
@@ -1,4 +1,0 @@
-
-from .freesurfer import freesurfer_main
-
-freesurfer_main()
